Extract home_dir into home crate
