+++ src/home/Cargo.toml
@@ -1,15 +1,11 @@
 [package]
-
-name = "rustup-utils"
-version = "1.5.0"
-authors = [ "Diggory Blake <diggsey@googlemail.com>" ]
-description = "Utility functions for rustup"
-
-documentation = "http://rust-lang-nursery.github.io/rustup.rs/rustup_utils/index.html"
-homepage = "https://github.com/rust-lang-nursery/rustup.rs"
-repository = "https://github.com/rust-lang-nursery/rustup.rs"
-
-license = "MIT OR Apache-2.0"
+name = "home"
+version = "0.1.0"
+authors = [ "Brian Anderson <andersrb@gmail.com>" ]
+description = "Shared definitions of home directories"
+documentation = "https://docs.rs/home"
+repository = "https://github.com/brson/home"
+license = "MIT/Apache-2.0"
 
 [dependencies]
 rand = "0.3.11"
@@ -23,7 +19,6 @@ toml = "0.1.27"
 semver = "0.4.0"
 download = { path = "../download" }
 lazy_static = "0.1.15"
-home = { path = "../home" }
 
 [target."cfg(windows)".dependencies]
 winapi = "0.2.8"
